requirements.txt
imapclient 
pyyaml